actions update

+++ src/actions/actions.ts
@@ -3,11 +3,11 @@
 import prisma from "@/lib/db";
 import { PetEssentials } from "@/lib/types";
 import { sleep } from "@/lib/utils";
-import { petFormSchema } from "@/lib/validations";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
 import { Pet } from "@prisma/client";
 import { revalidatePath } from "next/cache";
 
-export async function addPet(pet: PetEssentials) {
+export async function addPet(pet: unknown) {
   await sleep(1000);
 
   const validatedPet = petFormSchema.safeParse(pet)
@@ -29,13 +29,23 @@ export async function addPet(pet: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function editPet(petId: Pet["id"], newPetData: PetEssentials) {
+export async function editPet(petId: unknown, newPetData: unknown) {
+
+  const validatedId = petIdSchema.safeParse(petId);
+  
+  const validatedPet = petFormSchema.safeParse(newPetData)
+  if (!validatedId.success || !validatedPet.success) {
+    return {
+      message: "Invalid pet data."
+    }
+  }
+  
   try {
     await prisma.pet.update({
       where: {
-        id: petId,
+        id: validatedId.data,
       },
-      data: newPetData,
+      data: validatedPet.data,
     });
   } catch (error) {
     return {
@@ -45,11 +55,19 @@ export async function editPet(petId: Pet["id"], newPetData: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function deletePet(petId: Pet["id"]) {
+export async function deletePet(petId: unknown) {
+  const validatedId = petIdSchema.safeParse(petId);
+
+  if (!validatedId.success) {
+    return {
+      message: "Somthing went wrong."
+    }
+  }
+  
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedId.data,
       },
     });
   } catch (error) {
